requirements.txt
# Core dependencies
feedparser==6.0.11
requests==2.31.0
python-dateutil==2.8.2

# AI/ML
anthropic==0.39.0
openai==1.40.0

# NLP and text processing
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.3.0

# Database
sqlalchemy==2.0.31
alembic==1.13.2

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# CLI
rich==13.7.1
click==8.1.7

# Scheduling
schedule==1.2.2

# WhatsApp/SMS
twilio==9.3.7

# Testing
pytest==8.3.2
pytest-cov==5.0.0
